Extract port-checking logic into common module
