+++ integration-tests/redis-devservices/src/test/java/io/quarkus/redis/devservices/it/DevServicesRedisCustomPortReusableServiceITest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.redis.devservices.it.profiles.DevServicesCustomPortReusableServiceProfile;
-import io.quarkus.redis.devservices.it.utils.SocketKit;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
+import io.quarkus.test.ports.SocketKit;
 
 @QuarkusTest
 @TestProfile(DevServicesCustomPortReusableServiceProfile.class)
