+++ test-framework/junit5-internal/src/main/java/io/quarkus/test/ports/SocketKit.java
@@ -1,11 +1,11 @@
-package io.quarkus.redis.devservices.it.utils;
+package io.quarkus.test.ports;
 
 import java.io.IOException;
 import java.net.Socket;
 
 public class SocketKit {
 
-    public static boolean isPortAlreadyUsed(Integer port) {
+    public static boolean isPortAlreadyUsed(int port) {
         try (Socket ignored = new Socket("localhost", port)) {
             ignored.close();
             return true;
